Añado funciones para obtener json, modelo función para ver si una canción está en la base de datos, inserta en la base de datos y modifico la jerarquía de archivos
